+++ test/unit/fuzz-tools.test.js
@@ -479,6 +479,7 @@ import * as tegnerActivityV500 from '../../lib/tegner-activity-v500.js';
 import * as ludwigHairlossV501 from '../../lib/ludwig-hairloss-v501.js';
 import * as norwoodHairlossV502 from '../../lib/norwood-hairloss-v502.js';
 import * as simpsonMeningiomaV503 from '../../lib/simpson-meningioma-v503.js';
+import * as metavirFibrosisV504 from '../../lib/metavir-fibrosis-v504.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -903,6 +904,7 @@ const MODULES = {
   'ludwig-hairloss-v501.js': ludwigHairlossV501,
   'norwood-hairloss-v502.js': norwoodHairlossV502,
   'simpson-meningioma-v503.js': simpsonMeningiomaV503,
+  'metavir-fibrosis-v504.js': metavirFibrosisV504,
   'tb-testing.js': tbTesting,
 };